fix: service_id u32 overflow and Grafana curly-brace multi-select parsing Service IDs are u32 in JAM but stored as PostgreSQL INT (signed i32). Values > 2^31 wrapped to negative, causing hex formatting to produce "0x-15608d84" instead of "0xea9f727c" and parse_service_ids to silently drop large hex values. Fix by casting i32→u32 at the API boundary and parsing hex as u32. Also strip Grafana multi-select {a,b} curly braces from service and node filter params.

+++ src/grafana_store.rs
@@ -537,7 +537,7 @@ impl EventStore {
             .iter()
             .map(|row| {
                 serde_json::json!({
-                    "service_id": format!("0x{:x}", row.get::<i32, _>("service_id")),
+                    "service_id": format!("0x{:x}", row.get::<i32, _>("service_id") as u32),
                     "work_packages": row.get::<Option<i64>, _>("work_packages").unwrap_or(0),
                     "refinements": row.get::<Option<i64>, _>("refinements").unwrap_or(0),
                     "refinement_gas": row.get::<Option<i64>, _>("refinement_gas").unwrap_or(0),
@@ -604,7 +604,7 @@ impl EventStore {
             .map(|row| {
                 serde_json::json!({
                     "ts": row.get::<DateTime<Utc>, _>("ts"),
-                    "service_id": format!("0x{:x}", row.get::<i32, _>("service_id")),
+                    "service_id": format!("0x{:x}", row.get::<i32, _>("service_id") as u32),
                     "work_packages": row.get::<Option<i64>, _>("work_packages").unwrap_or(0),
                     "authorization_gas": row.get::<Option<i64>, _>("authorization_gas").unwrap_or(0),
                     "refinement_gas": row.get::<Option<i64>, _>("refinement_gas").unwrap_or(0),
